:hammer: Move ShellbotClient
